+++ rentACar/src/main/java/kodlama/io/rentACar/dto/requests/CreateModelRequest.java
@@ -16,5 +16,5 @@ public class CreateModelRequest {
     @Size(min = 2, max = 20)
     private String name;
     @NotNull
-    private int brandId;
+    private Long brandId;
 }
